feat!(array): move the array module to the private API
BREAKING CHANGE: this breaks user code that might be relying on
the lv_array module. The migration path for these users is to use
the private API by either including "lvgl_private.h" or enabling
LV_USE_PRIVATE_API
