+++ src/draw/lv_draw_vector.c
@@ -9,6 +9,7 @@
 #include "lv_draw_vector_private.h"
 #include "../misc/lv_area_private.h"
 #include "lv_draw_private.h"
+#include "../misc/lv_array.h"
 
 #if LV_USE_VECTOR_GRAPHIC
