+++ src/draw/vg_lite/lv_draw_vg_lite_vector.c
@@ -18,6 +18,7 @@
 #include "lv_vg_lite_grad.h"
 #include "lv_vg_lite_stroke.h"
 #include "../../image/lv_image_decoder_private.h"
+#include "../../misc/lv_array.h"
 #include "../lv_draw_vector_private.h"
 #include <float.h>
 #include <math.h>
