+++ src/draw/vg_lite/lv_vg_lite_pending.c
@@ -10,6 +10,7 @@
 #include "lv_vg_lite_pending.h"
 
 #if LV_USE_DRAW_VG_LITE
+#include "../../misc/lv_array.h"
 
 /*********************
  *      DEFINES
