+++ src/image/svg/lv_svg_render.c
@@ -14,6 +14,7 @@
 #include <math.h>
 #include <string.h>
 #include "lv_svg_render.h"
+#include "../../misc/lv_array.h"
 
 #if LV_USE_FREETYPE
     #include "../../font/freetype/lv_freetype_private.h"
