+++ src/image/svg/lv_svg_token.h
@@ -14,6 +14,8 @@
 
 #if LV_USE_SVG
 
+#include "../../misc/lv_array.h"
+
 /*********************
  *      DEFINES
  *********************/
