+++ src/misc/lv_circle_buf_private.h
@@ -16,6 +16,7 @@ extern "C" {
  *********************/
 
 #include "../lvgl_public.h"
+#include "lv_array.h"
 
 /*********************
  *      DEFINES
